+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Minor Improvements
 
 * PresenceValidator now returns false for strings that consist only of spaces
+* allows ShowEditPod to receive `false` in it's `onEdit` prop to skip rendering of the default edit icon
 
 # 0.21.0